TsukihiRenderer, PathtracingRenderer をクラスに昇格
