+++ tukihi.vcxproj.filters
@@ -1,12 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <ClInclude Include="pathtracing_radiance.h">
-      <Filter>Header Files\radiance</Filter>
-    </ClInclude>
-    <ClInclude Include="fake_radiance.h">
-      <Filter>Header Files\radiance</Filter>
-    </ClInclude>
     <ClInclude Include="constant.h">
       <Filter>Header Files</Filter>
     </ClInclude>
@@ -26,9 +20,6 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="scene.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="renderer.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="raymarching_menger_sponge.h">
@@ -73,13 +64,19 @@
     <ClInclude Include="raymarching_menger_sponge_loop.h">
       <Filter>Header Files\object</Filter>
     </ClInclude>
+    <ClInclude Include="renderer.h">
+      <Filter>Header Files\renderer</Filter>
+    </ClInclude>
+    <ClInclude Include="tsukihi_renderer.h">
+      <Filter>Header Files\renderer</Filter>
+    </ClInclude>
+    <ClInclude Include="pathtracing_renderer.h">
+      <Filter>Header Files\renderer</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Filter Include="Header Files">
       <UniqueIdentifier>{05afe2b4-690d-4c36-b49f-2f01b69eda20}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Header Files\radiance">
-      <UniqueIdentifier>{c574d642-3dcd-43f9-94e5-ab35bfbf9d2c}</UniqueIdentifier>
     </Filter>
     <Filter Include="Source Files">
       <UniqueIdentifier>{7bffbb7c-4614-460a-8886-d937dc8060c6}</UniqueIdentifier>
@@ -92,6 +89,9 @@
     </Filter>
     <Filter Include="Header Files\lib">
       <UniqueIdentifier>{328ac74f-b9a8-4fc2-80b0-5b329b72b274}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Header Files\renderer">
+      <UniqueIdentifier>{c574d642-3dcd-43f9-94e5-ab35bfbf9d2c}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
